requirements.txt
python-telegram-bot
pandas
pytz
openpyxl
httpx
Flask==2.0.1
pandas==1.3.3
openpyxl==3.0.9